pandoc-crossref: Source from GitHub instead of Hackage
As discussed in lierdakil/pandoc-crossref#360, Hackage metadata updates don't propagate into the source archives its hosts, so dependency bump versions don't create new versions that Homebrew's tooling can see. This is a running issue with pandoc-crossref, that updates necessary for it to keep up with pandoc releases causing some confusion (e.g. lierdakil/pandoc-crossref#367). #118436 pulled in b3c5f71a which should nudge a human to bump the revision field in the formulae, but that's inconvenient and requires a human.

@lierdakil [suggested][1] using the GitHub source archive instead of Hackage and [again][2]. I [suggested it][3], too, so let's do that.

The hash of the source archive is different because the Hackage archive doesn't have the metadata files that the GitHub archive has. Otherwise, the source is the same.

[1]: https://github.com/lierdakil/pandoc-crossref/issues/360#issuecomment-1289764832
[2]: https://github.com/lierdakil/pandoc-crossref/issues/367#issuecomment-1345630540
[3]: https://github.com/Homebrew/homebrew-core/issues/117921#issuecomment-1356827775